Refine example InfoBox panels
